Added Teams to team table

+++ Backend/AuthService/src/main/java/com/geolocation/authservice/data/models/RegisterUserDTO.java
@@ -29,7 +29,7 @@ public class RegisterUserDTO {
     private String phoneNumber;
 
     @NotNull(message = "Team Id is required")
-    private int teamId;
+    private long teamId;
 
     @NotNull(message = "Role must not be null")
     private RoleName role;
